melhor orgnizando rotas: login, logout & refazendo verificação de usuário em rotas protegidas
